Kill off JSONView and merge all JSON binding functionality into View.
