+++ Sources/ObjectivelyMVC/CollectionView.c
@@ -66,7 +66,7 @@ static void awakeWithDictionary(View *self, const Dictionary *dictionary) {
 		MakeInlet("itemSpacing", InletTypeSize, &this->itemSpacing, NULL)
 	);
 
-	$$(JSONView, applyInlets, self, dictionary, inlets);
+	$(self, bind, dictionary, inlets);
 }
 
 /**
