+++ Sources/ObjectivelyMVC/Panel.c
@@ -59,7 +59,7 @@ static void awakeWithDictionary(View *self, const Dictionary *dictionary) {
 		MakeInlet("isResiable", InletTypeBool, &this->isResizable, NULL)
 	);
 
-	$$(JSONView, applyInlets, self, dictionary, inlets);
+	$(self, bind, dictionary, inlets);
 }
 
 /**
