+++ Sources/ObjectivelyMVC/StackView.c
@@ -57,7 +57,7 @@ static void awakeWithDictionary(View *self, const Dictionary *dictionary) {
 		MakeInlet("spacing", InletTypeInteger, &this->spacing, NULL)
 	);
 
-	$$(JSONView, applyInlets, self, dictionary, inlets);
+	$(self, bind, dictionary, inlets);
 }
 
 /**
